projectApp/banking-App/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib[bcrypt]